eliminate doubled mouseevents
When a component is first poked, a synthetic no-button mouse event was
sent to the component, before the real event. This doesn't seem to serve
any purpose and was causing all components to get double mouse-pressed
events. In the case of DipSwitch and Button, this was causing
misbehavior. In all other cases, the second was ignored, but not needed.