DragDrop to toolbar works again
DragDrop was failing sporadically due to quirks in the way drag-and-drop
local transfer is handled. During a drag, before the drop, we can't rely
on local within-jvm transport of objects. So we need to pass enough
information in strings to make a determination about whether the drag
would be valid.